split up TestResultFactoryTests
