+++ TechTalk.SpecFlow/CucumberMessages/ITestResultFactory.cs
@@ -5,18 +5,6 @@ namespace TechTalk.SpecFlow.CucumberMessages
 {
     public interface ITestResultFactory
     {
-        IResult<TestResult> BuildPassedResult(ulong durationInNanoseconds);
-
-        IResult<TestResult> BuildFailedResult(ulong durationInNanoseconds, string message);
-
-        IResult<TestResult> BuildAmbiguousResult(ulong durationInNanoseconds, string message);
-
-        IResult<TestResult> BuildPendingResult(ulong durationInNanoseconds, string message);
-
-        IResult<TestResult> BuildSkippedResult(ulong durationInNanoseconds, string message);
-
-        IResult<TestResult> BuildUndefinedResult(ulong durationInNanoseconds, string message);
-
         IResult<TestResult> BuildFromContext(ScenarioContext scenarioContext, FeatureContext featureContext);
     }
 }
